slidesprodv2/requirements/prod.txt
djangae==0.9.10
django==1.10
git+https://github.com/mozilla/django-csp.git#egg=djangocsp
git+https://github.com/adamalton/django-csp-reports.git#egg=cspreports
django-session-csrf
six
#djangorestframework
#django-filter
# Uncomment these to use mapreduce or djangae.contrib.uniquetool (which uses mapreduce)
# These are forks of Google's libraries, which contain various fixes and patches
#git+https://github.com/potatolondon/appengine-pipelines.git
#git+https://github.com/potatolondon/potato-mapreduce.git#egg=mapreduce

# Uncomment this if you want to use djangae.fields.ComputedCollationField
# pyuca==1.1.2
